Streamlit_UI/requirements.txt
streamlit==1.12.2
Pillow==9.0.1
tensorflow-cpu==2.7.2
plotly==5.6.0
